MERN stack

+++ Other Courses/React/MERN Stack Tutorial/frontend/src/components/WorkoutDetails.jsx
@@ -1,14 +1,23 @@
 import React from 'react';
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+    
     const response = await fetch("/api/workouts/" + workout._id, {
-      method: "DELETE"
+      method: "DELETE",
+      headers: {
+        "Authorization": `Bearer ${user.token}`
+      }
     });
     const json = await response.json();
 
@@ -23,7 +32,10 @@ const WorkoutDetails = ({ workout }) => {
       <p><strong>Load (kg): </strong>{workout.load}</p>
       <p><strong>Reps: </strong>{workout.reps}</p>
       <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
-      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
+      <span
+        className="material-symbols-outlined"
+        onClick={handleClick}
+      >delete</span>
     </div>
   );
 };
